+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/ChatIdentifier.kt
@@ -97,6 +97,7 @@ object ChatIdentifierSerializer : KSerializer<ChatIdentifier> {
     override val descriptor: SerialDescriptor = internalSerializer.descriptor
     override fun deserialize(decoder: Decoder): ChatIdentifier {
         val id = internalSerializer.deserialize(decoder)
+
         return id.longOrNull ?.let {
             ChatId(it)
         } ?: id.content.let {
@@ -111,6 +112,44 @@ object ChatIdentifierSerializer : KSerializer<ChatIdentifier> {
     override fun serialize(encoder: Encoder, value: ChatIdentifier) {
         when (value) {
             is IdChatIdentifier -> encoder.encodeLong(value.chatId)
+            is Username -> encoder.encodeString(value.username)
+        }
+    }
+}
+
+@RiskFeature
+object FullChatIdentifierSerializer : KSerializer<ChatIdentifier> {
+    private val internalSerializer = JsonPrimitive.serializer()
+    override val descriptor: SerialDescriptor = internalSerializer.descriptor
+    override fun deserialize(decoder: Decoder): ChatIdentifier {
+        val id = internalSerializer.deserialize(decoder)
+
+        return id.longOrNull ?.let {
+            ChatId(it)
+        } ?:let {
+            val splitted = id.content.split("/")
+            if (splitted.size == 2) {
+                val (chatId, threadId) = splitted
+                ChatIdWithThreadId(
+                    chatId.toLongOrNull() ?: return@let null,
+                    threadId.toLongOrNull() ?: return@let null
+                )
+            } else {
+                null
+            }
+        } ?: id.content.let {
+            if (!it.startsWith("@")) {
+                Username("@$it")
+            } else {
+                Username(it)
+            }
+        }
+    }
+
+    override fun serialize(encoder: Encoder, value: ChatIdentifier) {
+        when (value) {
+            is ChatId -> encoder.encodeLong(value.chatId)
+            is ChatIdWithThreadId -> encoder.encodeString("${value.chatId}/${value.threadId}")
             is Username -> encoder.encodeString(value.username)
         }
     }