+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `Core`:
   * Add opportunity to create `IdChatIdentifier` with optional `threadId`
+  * New serializer `FullChatIdentifierSerializer` with serialization of `ChatIdWithThreadId`
 
 ## 4.1.0